tables.sql

+++ tables.sql
@@ -5,7 +5,7 @@ CREATE TABLE
         gears INT NOT NULL,
         brand VARCHAR(255) NOT NULL,
         model VARCHAR(255) NOT NULL,
-        price INT NOT NULL
+        price DECIMAL(10,2) NOT NULL
     );
 SELECT * FROM `Hubees_Shop`;
